plugins/connection/secure: fix 'writeboot' command with 'secure' port
When the secure-boot mode is enabled, therefore connecting to the secure
SAM-BA monitor and running secure applets to program some external NVM, we
should not patch the 6th vector of the ciphered bootstrap as we do with a
plain bootstrap. Indeed the vectors are now ciphered and the 6th vector
has already been patched before being ciphered by the secure SAM-BA
cipher.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>