plugsin/connection/serial: try to return to terminal mode before closing
SAM-BA cannot be used multiple times (without restarting the
processor) because it assumes that monitor is in terminal
mode. However, it places and leaves the monitor in normal mode on
exit.

This commit enables consecutive usage of sam-ba by placing the monitor
back into terminal mode when closing the serial connection.