sama5d2: Fix section numbers in datasheet references
According to revision history of the SAMA5D2 Series Datasheet, from
revision 11267E to DS60001476A a section “Safety and Security Features”
was added, so the numbering of later sections changed. The recent
revision is DS60001476C, references are updated to point to that one.

This was partly fixed with 65aab310d9f5 ("plugins/device/sama5d2: fix
datasheet section number for the NAND flash header"), but there are more
references on datasheet sections …

Signed-off-by: Alexander Dahl <ada@thorsis.com>